if changing from GDT_Int32 to GDT_Float64, needed to change the intermediate storage containers from int to double. Note that this is not what we are currently doing in readInputFile.cpp
Note that while the result of this looks good, the storage type of the original .lcp file was GDT_Int32 "https://landfire.gov/fuel/landscape". Also, luckily it looks like because the copied .lcp file retains the original storage type of GDT_Int32, using GDT_Float64 values still results in a .lcp file where gdalinfo returns that the type is still "Int16" for all 8 bands.

+++ src/surface_input_nodata_filler/surface_input_nodata_filler.cpp
@@ -136,7 +136,7 @@ void importBandData(GDALDataset* poDS, int bandNum, AsciiGrid<double>* ascii_gri
     ascii_grid->set_headerData(nC, nR, xL, yL, cS, nDV, nDV, prjStr);
 
     // read in the data for the specific band into the ascii grid, one scanline at a time
-    int* panScanline = new int[nC];
+    double* panScanline = new double[nC];
 
     for( int i = nR - 1; i >= 0; i-- )
     {
@@ -223,8 +223,8 @@ void writeBandDataIntStyle(GDALDataset* poDS, int bandNum, AsciiGrid<double>* as
 
     GDALRasterBand *poBand = poDS->GetRasterBand(bandNum);
 
-    int *padfScanline;
-    padfScanline = new int[nXSize];
+    double *padfScanline;
+    padfScanline = new double[nXSize];
 
     for(int i = nYSize-1; i >= 0; i--)
     {
